refactor: Remove v2 directory and filename prefixes
- Moved v2 components to their parent directories
- Updated import paths to reference non-v2 paths
- Kept UI components (@/components/v2/ui/*) as is for future migration
- Consolidated to single version of components after settings v2 flag removal
