+++ apps/studio.giselles.ai/app/(main)/settings/account/page.tsx
@@ -1,5 +1,40 @@
-import AccountSettingPageV2 from "./v2/page";
+import { Button } from "@/components/v2/ui/button";
+import { ToastProvider } from "@/packages/contexts/toast";
+import { fetchUserTeams } from "@/services/teams";
+import TeamCreation from "@/services/teams/components/team-creation";
+import { Card } from "../components/card";
+import UserTeams from "./user-teams";
 
 export default async function AccountSettingPage() {
-	return <AccountSettingPageV2 />;
+	const teams = await fetchUserTeams();
+
+	return (
+		<ToastProvider>
+			<div className="flex flex-col gap-[24px]">
+				<h3
+					className="text-primary-100 font-semibold text-[28px] leading-[28px] tracking-[-0.011em] font-hubot"
+					style={{ textShadow: "0px 0px 20px hsla(207, 100%, 48%, 1)" }}
+				>
+					Overview
+				</h3>
+				<div className="flex flex-col gap-y-[16px]">
+					<Card
+						title="Teams"
+						description="The teams that are associated with your Giselle account."
+						action={{
+							component: (
+								<div className="grid placeitems-center">
+									<TeamCreation>
+										<Button>Create New Team</Button>
+									</TeamCreation>
+								</div>
+							),
+						}}
+					>
+						<UserTeams teams={teams} />
+					</Card>
+				</div>
+			</div>
+		</ToastProvider>
+	);
 }
