+++ apps/studio.giselles.ai/app/(main)/settings/team/agent-time-charge.tsx
@@ -3,7 +3,7 @@ import { fetchCurrentTeam, isProPlan } from "@/services/teams";
 import {
 	AgentTimeUsageForFreePlan,
 	AgentTimeUsageForProPlan,
-} from "../components/v2/agent-time-usage";
+} from "./components/agent-time-usage";
 
 export async function AgentTimeCharge() {
 	const currentTeam = await fetchCurrentTeam();
