+++ apps/studio.giselles.ai/app/(main)/settings/team/agent-usage-table.tsx
@@ -1,4 +1,4 @@
-import { LocalDateTime } from "../components/local-date-time";
+import { LocalDateTime } from "./components/local-date-time";
 
 export type AgentActivity = {
 	agentId: string;
