+++ apps/studio.giselles.ai/app/(main)/settings/team/team-name.tsx
@@ -1,5 +1,5 @@
 import { fetchCurrentTeam } from "@/services/teams";
-import { TeamNameForm } from "../v2/team-name-form";
+import { TeamNameForm } from "./team-name-form";
 
 export async function TeamName() {
 	const currentTeam = await fetchCurrentTeam();
