+++ apps/studio.giselles.ai/services/teams/components/team-creation.tsx
@@ -2,7 +2,7 @@ import { getUser } from "@/lib/supabase";
 import { isEmailFromRoute06 } from "@/lib/utils";
 import { formatStripePrice, stripe } from "@/services/external/stripe";
 import invariant from "tiny-invariant";
-import { fetchUserTeams } from "../../fetch-user-teams";
+import { fetchUserTeams } from "../fetch-user-teams";
 import { TeamCreationForm } from "./team-creation-form";
 
 export default async function TeamCreation({
